Create MultiplesOf3And5.java
This is the first entry in the challenge. 

The goal is to find the numbers in the range of 0 to a given number that are either multiple of 3 or 5. After finding these numbers, add them up and return the sum as the answer. I used 10 as the given number which gives us 3, 5, 6, and 9. 3 + 5 + 6 + 9 = 23, so the program will return the "long" 23.